Refactor config data holders
