+++ core/org.wso2.carbon.tomcat.ext/src/main/java/org/wso2/carbon/tomcat/ext/valves/RequestCorrelationIdValve.java
@@ -31,6 +31,7 @@ import org.apache.log4j.MDC;
 import org.osgi.service.component.ComponentContext;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Deactivate;
+import org.wso2.carbon.tomcat.ext.internal.CorrelationConfigDataHolder;
 import org.wso2.carbon.tomcat.ext.service.ConfigurableCorrelationLogService;
 
 import javax.servlet.ServletException;
@@ -116,7 +117,7 @@ public class RequestCorrelationIdValve extends ValveBase {
             // Associates the generated Correlation ID Mapping to the request so that,
             // it will be available for any other asynchronous flows or threads spawned for this request
             request.setAttribute(CORRELATION_ID_MDC_REQUEST_ATTRIBUTE_NAME, associateToThreadMap);
-            if (ConfigurableCorrelationLogService.isEnable()) {
+            if (CorrelationConfigDataHolder.isEnable()) {
                 long currentTime = System.currentTimeMillis();
                 long timeTaken = currentTime - requestStartTime;
                 logRequestDetails(currentTime, timeTaken, CORRELATION_LOG_REQUEST_START, request);
@@ -126,7 +127,7 @@ public class RequestCorrelationIdValve extends ValveBase {
                 getNext().invoke(request, response);
             }
         } finally {
-            if (ConfigurableCorrelationLogService.isEnable()) {
+            if (CorrelationConfigDataHolder.isEnable()) {
                 long currentTime = System.currentTimeMillis();
                 long timeTaken = currentTime - requestStartTime;
                 logRequestDetails(currentTime, timeTaken, CORRELATION_LOG_REQUEST_END, request);
