+++ core/org.wso2.carbon.user.core/src/main/java/org/wso2/carbon/user/core/ldap/ConfigurableCorrelationLogService.java
@@ -22,14 +22,13 @@ package org.wso2.carbon.user.core.ldap;
 import org.osgi.service.component.annotations.Component;
 import org.wso2.carbon.logging.correlation.CorrelationLogConfigurable;
 import org.wso2.carbon.logging.correlation.bean.ImmutableCorrelationLogConfig;
+import org.wso2.carbon.user.core.internal.CorrelationConfigDataHolder;
 
 @Component(
         immediate = true,
         service = CorrelationLogConfigurable.class
 )
 public class ConfigurableCorrelationLogService implements CorrelationLogConfigurable {
-    private static boolean enable;
-
     @Override
     public String getName() {
         return "ldap";
@@ -37,10 +36,6 @@ public class ConfigurableCorrelationLogService implements CorrelationLogConfigur
 
     @Override
     public void onConfigure(ImmutableCorrelationLogConfig config) {
-        enable = config.isEnable();
-    }
-
-    public static boolean isEnable() {
-        return enable;
+        CorrelationConfigDataHolder.setEnable(config.isEnable());
     }
 }
